requirement.txt
fastapi
uvicorn
sentence-transformers
chromadb
pydantic
openreview-py
google-genai
google-generativeai
